admin

+++ resources/views/articles/show.blade.php
@@ -20,12 +20,12 @@
         <td>{{ $post->id }}</td>
         <td>{{ $post->body }}</td>
         <td> <a href="{{ url('/posts', $post->id) }}" class="btn btn-info">View this Post</a></td>
-       
+       @if(Auth::user()->id==$post->user_id || Auth::user()->role=='admin')
         <td><a href="{{  route('posts.edit', $post->id) }}" class="btn btn-primary">Edit Post</a></td>
         <td>{!! Form::open( [ 'method'  => 'delete', 'route' => [ 'posts.destroy', $post->id ] ])!!}
                     {{ Form::submit('Delete', ['class' => 'btn btn-danger']) }}
                     {{ Form::close() }}</td>
-      
+      @endif
       </tr>
       @endforeach
     </tbody>
